# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A Social platform for the creative at heart",
    "version": "1.0.0",
    "title": "Authors Haven",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "ninjas-ah-backend.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/auth/facebook": {
      "get": {
        "tags": ["auth"],
        "summary": "Register/login a facebook user",
        "description": "Register/login a facebook user and fetch the user profile",
        "operationId": "Facebook",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User successfully authenticated"
          },
          "201": {
            "description": "Account Successfully created"
          },
          "409": {
            "description": "User alredy exist"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/google": {
      "get": {
        "tags": ["auth"],
        "summary": "Register/login a google user",
        "description": "Register/login a google user and fetch the user profile",
        "operationId": "Google",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User successfully authenticated"
          },
          "201": {
            "description": "Account Successfully created"
          },
          "409": {
            "description": "User alredy exist"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/twitter": {
      "get": {
        "tags": ["auth"],
        "summary": "Register/login a twitter user",
        "description": "Register/login a twitter user and fetch the user profile",
        "operationId": "Twitter",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User successfully authenticated"
          },
          "201": {
            "description": "Account Successfully created"
          },
          "409": {
            "description": "User alredy exist"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": ["auth"],
        "summary": "Logs out current logged in user",
        "description": "Blacklist a token",
        "operationId": "LogoutUser",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User successfully logged out"
          },
          "401": {
            "description": "User not authorized to blacklist a given token"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "access-token",
      "in": "header"
    }
  },
...
}